grpclb, dns: pass balancer addresses via resolver.State
This change implements gRFC A26: gRPCLB Selection:
https://github.com/grpc/proposal/blob/master/A26-grpclb-selection.md

Note that addresses passed to gRPCLB through resolver.State.Attributes will not
automatically select the "grpclb" load balancing policy.  This will result in a
behavior change for users currently using the DNS name resolver and relying
upon that behavior.  Instead, a service config specifying grpclb must be used.
For example:

{
  "loadBalancingConfig": [
    {"grpclb": {}}
  ]
}

While this is a behavior change for gRPC-Go, it brings gRPC in line with the
intended implementation for gRPCLB support, and the behavior of the C and Java
gRPC implementations, and is considered a bug fix.

For now, and until support for the resolver.GRPCLB value is deleted, the
"grpclb" load balancing policy will still be selected automatically when
addresses with those types are produced by a name resolver.  This will prevent
custom resolvers from being impacted by this behavior change until they need to
be modified due to the deletion of this deprecated value.