testutils: Add a context parameter to the Receive() method.
This is a follow-up PR for one of the comments made during code review
of another PR, that the Receive() method in the testutils package should
accept a context instead of using a default timeout value.

Also, there seems to have been two copies of this library. Deleted the
one inside xds.

Also, cleaned up some tests in xds.