Introduce multi:// resolver: dial multiple hosts as one ClientConn
This allows setting up a ClientConn with multiple SubConns if there is no other resolver that groups them together. I used this for example with my localhost Raft cluster, using health checks to talk to the leader.

This functionality is currently available as https://github.com/Jille/grpc-multi-resolver, but that needs a copy of ParseTarget() which is only available in internal/.