trim the Makefile
Remove sections which deal with `deps`. These are not required with the
move to Go modules, and create unnecessary diffs in go.mod and go.sum
files.

Left the rest as is, as some are used from `testing.yaml` and others
serve as a handy reference when one wants to build/test something
specific.