server: unwrap status errors from RPC handler methods.
Currently, if an RPC handler method returns an error with a GRPCStatus
method, the GRPC server calls that method and uses the result's code
and message in the RPC response. However, if the error in question
wraps an error with a GRPCStatus method, the RPC response has code
UNKNOWN. This commit changes status.FromError and status.Code to
also extract statuses from wrapped errors.

The use case is something like this: I have RPC server X; it has a
method that makes calls to RPC servers Y and Z, computes something
from the responses, and returns it. If there's an error talking to Y
or Z, then X's error response should have the same code as Y's or Z's
response did. Currently, I can accomplish this by returning the error
from Y or Z directly.

However, if there's some helpful code in the middle that wraps the
error from Y or Z (e.g. to say which of Y or Z goofed), then I have to
unwrap the error myself. This might not seem too onerous, but it means
that either (a) I need some boilerplate helper unwrapping for every
RPC handler method, or (b) any code between my RPC handler method and
the RPC stubs for Y or Z can break things at any time by wrapping an
error.

With this commit, the GRPC server handles the error unwrapping for me.

Error wrapping and unwrwapping were introduced in Go 1.13[1], so the
last three major releases of Go all have them.

[1] https://golang.org/doc/go1.13#error_wrapping