Add deadline on Read() calls as part of the client handshake.
In theory Go has a concurrent handshake limit of 100. However, if all the handshake attempts would get stuck in Read(), which may happen when the LOAS credentials are not available, we would eventually reach a deadlock. By having a deadline on Read() RPCs we guarantee that eventually the RPC would be cancelled leaving room for a new handshake attempt, which by the time LOAS credentials are available, it should succeed.