[eds_balancer_config_builder] cluster_resolver: add functions to build child balancer config
This change adds a package with functions to build a tree of balancer
configs.

The functions are designed for cluster_resolver, and can be used by the
EDS balancer (because EDS is a subset of cluster_resolver). For the same
reason, the functions are exported in a subpackage of cluster_resolver.