xds_resolver: fix flaky Test/XDSResolverDelayedOnCommitted (#4393)
Before this change, if two xds client updates came too close together, the second one could replace the first one.  The fix is to wait for the effects of the first update before sending the second update.  I injected a synthetic delay into handling the updates from the channel to reproduce this flake 100%, and confirmed this change fixes it.

As part of this change I also noticed that we're actually calling the context cancellation function twice via defers, and never the cancel function from the test setup, so I fixed that, too.