Prevent hang in Test/ClientUpdatesParamsAfterGoAway
In cases where `t.Fatalf()` doesn't completely end the program, a `ctx.Err()` will cause the program to wait on a mutex indefinitely. 

Forcing the test to enter this specific `if` condition should reproduce the issue.

Adding an unlock before reporting the error allows the program to fail the test as intended when a network error occurs.