internal/transport: do not mask ConnectionError
PR #4316 updates the http2Client `Close` method so it can append
additional debug information about prior GOAWAY messages before this
connection was closed.

However, it is not safe to do this in the general case, as the original
error is propagated through the connection's `Stream` via a `recvMsg`,
and read later on. If the additional metadata is appended by wrapping
the error with `fmt.Errorf`, downstream clients will not be able to
detect meaningful error conditions (such as `io.EOF`) or the temporarity
of the error.

Fix this by properly wrapping the original error with `connectionErrorf`
like it is done for all the other errors in the client.