[clusterimpl_pick_sc_wrong_type] xds/clusterimpl: fix SubConn wrapper returned by picker during race
The previous logic is that the picker updated with a non-Ready state
should always return error, so the clusterimpl picker doesn't try to
unwrap the SubConn.

The assumption is not true, in cases like ring_hash (where the picker is
mutable, and could start to return SubConns before the overall state is
updated to Ready during races).