grpc: enhance client interceptor documentation
Explain that only one interceptor may be attached with the
WithUnaryInterceptor and WithStreamInterceptor DialOptions and that
their respective Chain DialOptions should be used if more than one is
desired. Furthermore, explain what happens if either of the
aforementioned DialOptions are passed more than once to a dial call.
The reason for doing this is that in the course of addressing whether
interceptors would suffice for #5055, it was unclear what happened if
multiple were attached.

Provide a snippet for what an effectively trivial interceptor would be
in the documentation.  The motivation similarly is to illuminate a path
for users.