Support setting dial options for the xDS client.
This feature can be used to make sure the xDS client behavior can be
controlled to a certain extent from users.

The xdsclient package now exposes 3 different experimental APIs
1. SetControllerDialOpts which sets the extra dial options that the xds
controller uses.
2. ControllerDialOpts to return the dial options that have been set.
3. ResetControllerDialOpts to reset the dial options that might have
been set.

For Set/Reset to take effect they must be called before the xds client
is called. As xDS client is a singleton by default, this indicates that
the Set/Reset must ideally be called as part of a init function to make
sure any xDS client is created with the required dial options.

If the Set/Reset is called after the xDS client is created, they are
no-ops until the xDS client is recreated again.