Support setting dial interceptor for the xDS controller. This feature can be used control xDS client behavior through a Dial interceptor set from users.
Users can SetDialInterceptor on the controller package to intercept the
grpc.Dial call and inject/modify the target or the options.

ResetDialInterceptor can be used to reset and set the dialler back to
grpc.Dial.

Both of these calls will only take affect if they are called before the
controller is created using New.