credentials/google: stub out the oauth package in test
The oauth package, as part of checking whether it is running on GCE,
creates a persistent connection to the metadata server. This persistent
connection spawns a couple of goroutines, one to read and one to write.
And these goroutines show up in our leakchecker.

We are not testing perRPC credentials in this test and can safely stub
out the oauth package to make the leakchecker happy.