Support xDS credentials plugin.
Currently the credentials used for communication with the xDS management
server is restricted to "google_default" and "insecure". There is no way
for users to inject custom credentials, even though the bootstrap can be
used to pass in different values.

This PR adds support for a new credentials package which can be used to
register supported credentials along with a name that can be recognized
by the bootstrap.

The credential package also registers the currently supported
"google_default" and "insecure" options to maintain the status quo. To keep the
interface of the credentials builder clean, the insecure credentials now
satisfies the grpc credentials bundle interface.