rls: make the data cache purge ticker a field in rlsBalancer
The race detector was complaining about a data race on the
`periodicCachePurgeFreq` var which was being overridden in tests. This
commit changes things around a little bit by putting the ticker inside
the rlsBalancer and constructing it at build time. Since I have been
unable to reproduce this data race locally, hoping that this will help.