transport/http2_server: Make it possible to configure GOAWAY thresholds
As discussed in https://github.com/grpc/grpc/issues/25713, it's very
easy to wreak havoc by tweaking the gRPC keepalive interval on the
client. Previously the limit of 2 pings within 2 hours was hard-coded;
if a client went over this limit, the server sends a GOAWAY message is
sent and closes the connection.

https://github.com/grpc/grpc/blob/master/doc/keepalive.md recommends
that `GRPC_ARG_HTTP2_MAX_PINGS_WITHOUT_DATA` should be set to 0, and
that enforcement of this policy will be deprecated in the future. This
commit allows disabling of this enforcement.

Closes https://github.com/grpc/grpc-go/issues/5161