reflection: use protobuf/reflect instead of go reflection, fix dynamicpb
The reflection service panics when it encounters a dynamically-created message.

The panic happens here:

  m, ok := reflect.Zero(reflect.PtrTo(st)).Interface().(protoMessage)

This is because `st` is a `dynamicpb.Message` which can't be created using go reflection.

Rather than add special handling for dynamic messages, migrating the reflection service to use protobuf/reflection instead of go reflection both solves the problem and cleans up a lot of code.