benchmark: use default buffer sizes
The read and write buffer sizes are currently hardcoded to 128KB instead of the default 32KB. This
can potentially lead to misleading benchmark results in the most common case, since most users
presumably use the default value.

This change removes the custom value for read and write transport buffer sizes, both client and
server side, so that benchmarks use the default values instead.