encoding: clarify when it is safe to register compressors and codecs
Previously the docs suggested that it was only safe to register
compressors and codecs inide init() functions, I presume because
these encoding.RegisterCompressor and encoding.RegisterCodec are
not thread-safe. If that is the case, then it should also be safe
to register codecs in the main goroutine before any other
goroutines have been spawned.

Registering compressors and codecs from the main goroutine can
be useful, because it allows application developers to ensure
that their own compressor/codec is being used, and that
transitively imported packages haven't overridden their preferred
implementation.