client: fix ClientStream.Header() behavior
This change causes us to reliably report the stream status from
ClientStream.Header, including io.EOF in case the stream ended successfully
with a trailers-only response.  To do this required moving the binary logging
of trailers into clientStream.finish.

By doing this, retry is also fixed in cases where an error is encountered on
the stream before calling Header().  Previously, if Header() was called first,
we would report ErrNoHeaders from the transport which would be converted into a
non-error by clientStream.Header, which would then commit the RPC and prevent
retries.  A test case was added to cover this scenario.