transport: add callback to signal completion of writes
When writing data on an HTTP writer, the call completes once the frame
is added to the control buffer queue. Therefore, there is no way to know
when the actual write is completed, as it occurs asynchronously
afterward.

`transport.Options` now allows passing a callback that's called exactly
once all the data has been committed to the underlying http2 layer.