metadata: Use strings.EqualFold for ValueFromIncomingContext
This is approximately 30% faster in a microbenchmark I added. The
previous version using strings.ToLower() looks at every key byte.
This version will typically evaluate "not equal" on the first byte.

The test and benchmark had no coverage for this loop, so I added it.
I also revised the comments in FromIncomingContext and
ValueFromIncomingContext. The problem is not that the MD is created
using the helper functions. The MD must be attached to the context
using NewIncomingContext, which could lower case the keys. The
problem is that NewIncomingContext does not make a copy of its
argument, so callers can mutate it after.

This change adds a test for FromIncomingContext that tests this case,
and modifies the existing TestValueFromIncomingContext to test it.

Benchstat output from 10 runs of the benchmark on my laptop:

goos: darwin
goarch: arm64
pkg: google.golang.org/grpc/metadata
                                          │ before.txt  │              after.txt              │
                                          │   sec/op    │   sec/op     vs base                │
ValueFromIncomingContext/key-found-10       26.56n ± 2%   26.74n ± 2%        ~ (p=0.078 n=10)
ValueFromIncomingContext/key-not-found-10   50.48n ± 1%   36.23n ± 2%  -28.23% (p=0.000 n=10)
geomean                                     36.62n        31.13n       -15.00%