ringhash: fix bug where ring hash can be stuck in transient failure
Fixes #7363, which has a thorough problem description. The test cases added
reproduce the error without the fix, but not reliably, since the ring may be
constructed in a way where we don't end up getting stuck. Running the test
multiple times definitely end up triggering the problem.

The solution is to keep a separate list of available addresses, and when there
are no picks and we trigger connection attempts, try them one at a time.

RELEASE NOTES:

    ringhash: fix bug that could prevent the balancer to recover from transient failure.