server.go: Only call FromIncomingContext with stats handlers
Skip calling metadata.FromIncomingContext if there are no stats
handlers. This function is relatively expensive because it makes a
copy of the map of gRPC headers. For our production services that use
tracing libraries, this means there are typically about ~10 headers
on every call. For services with cheap gRPC calls, this can use
nearly 1% of CPU.

There are unit tests that cover this code path in the stats package
but sadly I did not find any existing microbenchmarks that we could
easily use to verify this improvement.