feat: export context errors
Since #6862 was closed as wontfix, this is a follow-up proposal to actuall pass along errors the caller may be interested it.

The current wrapped errors are only used for the specific cases of context cancellation and deadline. The rationale being: these are the errors that are caused via passed in context arguments from the caller.