transport: writeStatus: Skip Status.Proto() without details
Add Status.HasDetails() to check for details. Change
http2server.writeStatus to check to avoid unnecessary clones in the
common case when there are no details. This seems to be worth about
0.1% of CPU for a high QPS gRPC service, according to the Go
profiler.

While looking at profiles for a high QPS gRPC service at Datadog, I
noticed that about 0.3% of CPU time is being spent calling
Status.Proto(). I believe this is unnecessary: this clones the
protobuf status message, only to check if there are details. If there
are not (the common case, e.g. for status code OK), it throws the
protobuf message away.

I tried quickly but was not able to reliably measure the impact of
this change from any end-to-end benchmarks, but it does seem to make
this function go away according to the profiler. This should be a
small win for every gRPC server.