Fix panic in least_request balancer
There was a leftover embedded balancer in least request that was never set in
the code when it got refactored to pick first as leaf. As a result, method calls
forwarded to the embedded balancer would panic.

Remove the embedded balance and properly forward the missing methods.

Fixes https://github.com/grpc/grpc-go/issues/8332.