examples/features/health: Clarify magic import for clients
The google.golang.org/grpc/health package must be imported for
client health checking to work. I somehow missed this, even though
it is in the README, the client example, and the health package docs.
Attempt to make it clearer with a few extra mentions, since it is
quite hard to debug this misconfiguration.

* Remove deprecated grpc.WithBlock function
* Make service config const since it isn't modified

Attempts to clarify Issue 8590.