xdsclient: fix TestConcurrentReportLoad to not run for 10s (#8598)
While working on the fix for the xDS client unsubscribe/resubscribe
race, I noticed that the tests in the `internal/xds/xdsclient/tests/`
directory were taking about a minute to run. Upon inspection I found
that `TestConcurrentReportLoad` was running for the configured test
timeout duration of `10s`, but was not failing.

This PR fixes the test to run in a short duration. It also makes a
couple of other cleanups that I noticed when fixing this test.

RELEASE NOTES: none

---------

Co-authored-by: eshitachandwani <59800922+eshitachandwani@users.noreply.github.com>